Split more modules
