+++ snuba/api/query.py
@@ -8,10 +8,10 @@ from typing import Any, MutableMapping, NamedTuple
 from snuba import settings, state
 from snuba.clickhouse.native import ClickhousePool
 from snuba.clickhouse.query import ClickhouseQuery
+from snuba.query.columns import all_referenced_columns
 from snuba.request import Request
 from snuba.state.rate_limit import RateLimitAggregator, RateLimitExceeded, PROJECT_RATE_LIMIT_NAME
 from snuba.util import (
-    all_referenced_columns,
     create_metrics,
     force_bytes,
     Timer,
