+++ snuba/clickhouse/query.py
@@ -2,7 +2,11 @@ from typing import Sequence
 
 from snuba import settings as snuba_settings
 from snuba import util
-from snuba.datasets import Dataset
+from snuba.query.columns import (
+    column_expr,
+    conditions_expr,
+)
+from snuba.datasets.dataset import Dataset
 from snuba.query.parsing import ParsingContext
 from snuba.query.query import Query
 from snuba.request.request_settings import RequestSettings
@@ -27,11 +31,11 @@ class ClickhouseQuery:
 
         parsing_context = ParsingContext()
 
-        aggregate_exprs = [util.column_expr(dataset, col, query, parsing_context, alias, agg) for (agg, col, alias) in query.get_aggregations()]
+        aggregate_exprs = [column_expr(dataset, col, query, parsing_context, alias, agg) for (agg, col, alias) in query.get_aggregations()]
         groupby = util.to_list(query.get_groupby())
-        group_exprs = [util.column_expr(dataset, gb, query, parsing_context) for gb in groupby]
+        group_exprs = [column_expr(dataset, gb, query, parsing_context) for gb in groupby]
         column_names = query.get_selected_columns() or []
-        selected_cols = [util.column_expr(dataset, util.tuplify(colname), query, parsing_context) for colname in column_names]
+        selected_cols = [column_expr(dataset, util.tuplify(colname), query, parsing_context) for colname in column_names]
         select_clause = u'SELECT {}'.format(', '.join(group_exprs + aggregate_exprs + selected_cols))
 
         from_clause = u'FROM {}'.format(source)
@@ -55,15 +59,15 @@ class ClickhouseQuery:
 
         where_clause = ''
         if query.get_conditions():
-            where_clause = u'WHERE {}'.format(util.conditions_expr(dataset, query.get_conditions(), query, parsing_context))
+            where_clause = u'WHERE {}'.format(conditions_expr(dataset, query.get_conditions(), query, parsing_context))
 
         prewhere_clause = ''
         if prewhere_conditions:
-            prewhere_clause = u'PREWHERE {}'.format(util.conditions_expr(dataset, prewhere_conditions, query, parsing_context))
+            prewhere_clause = u'PREWHERE {}'.format(conditions_expr(dataset, prewhere_conditions, query, parsing_context))
 
         group_clause = ''
         if groupby:
-            group_clause = 'GROUP BY ({})'.format(', '.join(util.column_expr(dataset, gb, query, parsing_context) for gb in groupby))
+            group_clause = 'GROUP BY ({})'.format(', '.join(column_expr(dataset, gb, query, parsing_context) for gb in groupby))
             if query.has_totals():
                 group_clause = '{} WITH TOTALS'.format(group_clause)
 
@@ -71,11 +75,11 @@ class ClickhouseQuery:
         having_conditions = query.get_having()
         if having_conditions:
             assert groupby, 'found HAVING clause with no GROUP BY'
-            having_clause = u'HAVING {}'.format(util.conditions_expr(dataset, having_conditions, query, parsing_context))
+            having_clause = u'HAVING {}'.format(conditions_expr(dataset, having_conditions, query, parsing_context))
 
         order_clause = ''
         if query.get_orderby():
-            orderby = [util.column_expr(dataset, util.tuplify(ob), query, parsing_context) for ob in util.to_list(query.get_orderby())]
+            orderby = [column_expr(dataset, util.tuplify(ob), query, parsing_context) for ob in util.to_list(query.get_orderby())]
             orderby = [u'{} {}'.format(ob.lstrip('-'), 'DESC' if ob.startswith('-') else 'ASC') for ob in orderby]
             order_clause = u'ORDER BY {}'.format(', '.join(orderby))
